LFS-743: As an API consumer, I can access the human-readable answers in addition to the answer value in the form JSON Moved all the label processors in a separate package
