+++ modules/data-entry/src/main/java/ca/sickkids/ccm/lfs/dataentry/internal/serialize/labels/AnswerOptionsLabelProcessor.java
@@ -16,13 +16,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package ca.sickkids.ccm.lfs.dataentry.internal.serialize;
+package ca.sickkids.ccm.lfs.dataentry.internal.serialize.labels;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.LinkedHashMap;
+import java.util.Map;
 import java.util.function.Function;
 
 import javax.jcr.Node;
+import javax.jcr.NodeIterator;
+import javax.jcr.Property;
 import javax.jcr.RepositoryException;
 import javax.jcr.Value;
 import javax.json.Json;
@@ -35,12 +37,12 @@ import org.osgi.service.component.annotations.Component;
 import ca.sickkids.ccm.lfs.serialize.spi.ResourceJsonProcessor;
 
 /**
- * Gets the file name of the file question answer.
+ * Gets the human-readable question answer for text questions with options.
  *
  * @version $Id$
  */
 @Component(immediate = true)
-public class FileLabelProcessor extends SimpleAnswerLabelProcessor implements ResourceJsonProcessor
+public class AnswerOptionsLabelProcessor extends SimpleAnswerLabelProcessor implements ResourceJsonProcessor
 {
     @Override
     public String getName()
@@ -71,7 +73,7 @@ public class FileLabelProcessor extends SimpleAnswerLabelProcessor implements Re
         final Function<Node, JsonValue> serializeNode)
     {
         try {
-            if (child.isNodeType("lfs:FileResourceAnswer") || node.isNodeType("lfs:SomaticVariantsAnswer")) {
+            if (child.isNodeType("lfs:TextAnswer")) {
                 return serializeNode.apply(child);
             }
         } catch (RepositoryException e) {
@@ -84,20 +86,8 @@ public class FileLabelProcessor extends SimpleAnswerLabelProcessor implements Re
     public void leave(Node node, JsonObjectBuilder json, Function<Node, JsonValue> serializeNode)
     {
         try {
-            if (node.isNodeType("lfs:FileResourceAnswer") || node.isNodeType("lfs:SomaticVariantsAnswer")) {
+            if (node.isNodeType("lfs:TextAnswer")) {
                 addProperty(node, json, serializeNode);
-            }
-        } catch (RepositoryException e) {
-            // Really shouldn't happen
-        }
-    }
-
-    @Override
-    public void addProperty(Node node, JsonObjectBuilder json, Function<Node, JsonValue> serializeNode)
-    {
-        try {
-            if (node.hasProperty("value")) {
-                json.add(PROP_DISPLAYED_VALUE, getAnswerLabel(node, null));
             }
         } catch (RepositoryException e) {
             // Really shouldn't happen
@@ -108,22 +98,60 @@ public class FileLabelProcessor extends SimpleAnswerLabelProcessor implements Re
     public JsonValue getAnswerLabel(final Node node, final Node question)
     {
         try {
-            List<String> names = new ArrayList<>();
-            String fullPath = node.getPath() + "/";
-            Value[] childNodes = node.getProperty("value").getValues();
-            for (Value item : childNodes) {
-                String fileName = item.getString().replace(fullPath, "");
-                names.add(fileName);
+            Map<String, String> propsMap = new LinkedHashMap<>();
+
+            Property nodeProp = node.getProperty(PROP_VALUE);
+            if (nodeProp.isMultiple()) {
+                for (Value value : nodeProp.getValues()) {
+                    propsMap.put(value.getString(), value.getString());
+                }
+            } else {
+                propsMap.put(nodeProp.getString(), nodeProp.getString());
             }
 
-            if (names.size() == 1) {
-                return Json.createValue(names.get(0));
+            if (question == null) {
+                return createJsonArrayFromList(propsMap.values());
             }
 
-            return createJsonArrayFromList(names);
-        } catch (final RepositoryException ex) {
+            processOptions(question, propsMap);
+
+            if (propsMap.size() == 1) {
+                return Json.createValue((String) propsMap.values().toArray()[0]);
+            }
+
+            return createJsonArrayFromList(propsMap.values());
+        } catch (RepositoryException e) {
             // Really shouldn't happen
         }
         return null;
+    }
+
+    private void processOptions(final Node question, Map<String, String> propsMap)
+    {
+        try {
+            NodeIterator childNodes = question.getNodes();
+            if (childNodes.getSize() > 0) {
+                int count = 0;
+                while (childNodes.hasNext()) {
+                    Node optionNode = childNodes.nextNode();
+                    if (!"lfs:AnswerOption".equals(optionNode.getPrimaryNodeType().getName())
+                            || !optionNode.hasProperty(PROP_VALUE)) {
+                        continue;
+                    }
+
+                    String option = optionNode.getProperty(PROP_VALUE).getString();
+                    if (propsMap.containsKey(option) && optionNode.hasProperty(PROP_LABEL)) {
+                        propsMap.put(option, optionNode.getProperty(PROP_LABEL).getString());
+                        count++;
+                    }
+
+                    if (propsMap.size() == count) {
+                        break;
+                    }
+                }
+            }
+        } catch (RepositoryException e) {
+            // Really shouldn't happen
+        }
     }
 }
