+++ modules/data-entry/src/main/java/ca/sickkids/ccm/lfs/dataentry/internal/serialize/labels/FileLabelProcessor.java
@@ -16,15 +16,13 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package ca.sickkids.ccm.lfs.dataentry.internal.serialize;
+package ca.sickkids.ccm.lfs.dataentry.internal.serialize.labels;
 
-import java.util.LinkedHashMap;
-import java.util.Map;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.function.Function;
 
 import javax.jcr.Node;
-import javax.jcr.NodeIterator;
-import javax.jcr.Property;
 import javax.jcr.RepositoryException;
 import javax.jcr.Value;
 import javax.json.Json;
@@ -37,12 +35,12 @@ import org.osgi.service.component.annotations.Component;
 import ca.sickkids.ccm.lfs.serialize.spi.ResourceJsonProcessor;
 
 /**
- * Gets the human-readable question answer for text questions with options.
+ * Gets the file name of the file question answer.
  *
  * @version $Id$
  */
 @Component(immediate = true)
-public class AnswerOptionsLabelProcessor extends SimpleAnswerLabelProcessor implements ResourceJsonProcessor
+public class FileLabelProcessor extends SimpleAnswerLabelProcessor implements ResourceJsonProcessor
 {
     @Override
     public String getName()
@@ -73,7 +71,7 @@ public class AnswerOptionsLabelProcessor extends SimpleAnswerLabelProcessor impl
         final Function<Node, JsonValue> serializeNode)
     {
         try {
-            if (child.isNodeType("lfs:TextAnswer")) {
+            if (child.isNodeType("lfs:FileResourceAnswer") || node.isNodeType("lfs:SomaticVariantsAnswer")) {
                 return serializeNode.apply(child);
             }
         } catch (RepositoryException e) {
@@ -86,8 +84,20 @@ public class AnswerOptionsLabelProcessor extends SimpleAnswerLabelProcessor impl
     public void leave(Node node, JsonObjectBuilder json, Function<Node, JsonValue> serializeNode)
     {
         try {
-            if (node.isNodeType("lfs:TextAnswer")) {
+            if (node.isNodeType("lfs:FileResourceAnswer") || node.isNodeType("lfs:SomaticVariantsAnswer")) {
                 addProperty(node, json, serializeNode);
+            }
+        } catch (RepositoryException e) {
+            // Really shouldn't happen
+        }
+    }
+
+    @Override
+    public void addProperty(Node node, JsonObjectBuilder json, Function<Node, JsonValue> serializeNode)
+    {
+        try {
+            if (node.hasProperty("value")) {
+                json.add(PROP_DISPLAYED_VALUE, getAnswerLabel(node, null));
             }
         } catch (RepositoryException e) {
             // Really shouldn't happen
@@ -98,60 +108,22 @@ public class AnswerOptionsLabelProcessor extends SimpleAnswerLabelProcessor impl
     public JsonValue getAnswerLabel(final Node node, final Node question)
     {
         try {
-            Map<String, String> propsMap = new LinkedHashMap<>();
-
-            Property nodeProp = node.getProperty(PROP_VALUE);
-            if (nodeProp.isMultiple()) {
-                for (Value value : nodeProp.getValues()) {
-                    propsMap.put(value.getString(), value.getString());
-                }
-            } else {
-                propsMap.put(nodeProp.getString(), nodeProp.getString());
+            List<String> names = new ArrayList<>();
+            String fullPath = node.getPath() + "/";
+            Value[] childNodes = node.getProperty("value").getValues();
+            for (Value item : childNodes) {
+                String fileName = item.getString().replace(fullPath, "");
+                names.add(fileName);
             }
 
-            if (question == null) {
-                return createJsonArrayFromList(propsMap.values());
+            if (names.size() == 1) {
+                return Json.createValue(names.get(0));
             }
 
-            processOptions(question, propsMap);
-
-            if (propsMap.size() == 1) {
-                return Json.createValue((String) propsMap.values().toArray()[0]);
-            }
-
-            return createJsonArrayFromList(propsMap.values());
-        } catch (RepositoryException e) {
+            return createJsonArrayFromList(names);
+        } catch (final RepositoryException ex) {
             // Really shouldn't happen
         }
         return null;
-    }
-
-    private void processOptions(final Node question, Map<String, String> propsMap)
-    {
-        try {
-            NodeIterator childNodes = question.getNodes();
-            if (childNodes.getSize() > 0) {
-                int count = 0;
-                while (childNodes.hasNext()) {
-                    Node optionNode = childNodes.nextNode();
-                    if (!"lfs:AnswerOption".equals(optionNode.getPrimaryNodeType().getName())
-                            || !optionNode.hasProperty(PROP_VALUE)) {
-                        continue;
-                    }
-
-                    String option = optionNode.getProperty(PROP_VALUE).getString();
-                    if (propsMap.containsKey(option) && optionNode.hasProperty(PROP_LABEL)) {
-                        propsMap.put(option, optionNode.getProperty(PROP_LABEL).getString());
-                        count++;
-                    }
-
-                    if (propsMap.size() == count) {
-                        break;
-                    }
-                }
-            }
-        } catch (RepositoryException e) {
-            // Really shouldn't happen
-        }
     }
 }
